src/requirements.txt
ansible==2.8.1
asn1crypto==0.24.0
backports.csv==1.0.7
cffi==1.12.3
cryptography==2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-bootstrap4==0.0.8
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==2.6.2
pycparser==2.19
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
uwsgi==2.0.15
